Install assets on first run
